Remove redux tests and move helpers directory
